feat: 취미 경력 대신 취미 시작 년월을 받도록 변경

+++ lime-domain/src/main/java/com/programmers/lime/domains/member/model/MemberProfile.java
@@ -11,7 +11,7 @@ public record MemberProfile(
 	String profileImage,
 	int level,
 	String hobby,
-	int career,
+	long career,
 	String favorability,
 	String content,
 	String mbti
